requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
python-dateutil==2.8.2
